refactor(agents-desktop): extract settings and API key helpers
Move settings serialization and API key persistence into focused modules so main.ts can keep only desktop orchestration.
